+++ src/channels.js
@@ -1,0 +1,3 @@
+export function defineChannel(definition) {
+  return definition;
+}