+++ src/connections.js
@@ -1,0 +1,3 @@
+export function defineConnection(definition) {
+  return definition;
+}